requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
twilio>=8.0.0
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
itsdangerous>=2.1.0
celery>=5.3.0
redis>=5.0.0
pytest>=7.4.0
httpx>=0.24.0
psycopg2-binary>=2.9.0
alembic>=1.11.0


